move wallet tests that are timing out on CI to slow tests (#4415)
